reorganize fonll commands
