+++ src/pineko/cli/_base.py
@@ -1,6 +1,10 @@
 """Adds global CLI options."""
 
+import pathlib
+
 import click
+
+from .. import configs
 
 CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
 
@@ -8,3 +12,22 @@ CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
 @click.group(context_settings=CONTEXT_SETTINGS)
 def command():
     """pineko: Combine PineAPPL grids and EKOs into FK tables."""
+
+
+config_option = click.option(
+    "-c",
+    "--configs",
+    "cfg",
+    default=None,
+    type=click.Path(resolve_path=True, path_type=pathlib.Path),
+    help="Explicitly specify config file (it has to be a valid TOML file).",
+)
+
+
+def load_config(cfg):
+    """Load configuration files."""
+    path = configs.detect(cfg)
+    base_configs = configs.load(path)
+    configs.configs = configs.defaults(base_configs)
+    if cfg is not None:
+        print(f"Configurations loaded from '{path}'")
