+++ src/pineko/cli/scaffold.py
@@ -1,30 +1,17 @@
 """'scaffold' mode of CLI."""
 
-import pathlib
-
 import click
 import rich
 
 from .. import configs, scaffold
-from ._base import command
+from ._base import command, config_option, load_config
 
 
 @command.group("scaffold")
-@click.option(
-    "-c",
-    "--configs",
-    "cfg",
-    default=None,
-    type=click.Path(resolve_path=True, path_type=pathlib.Path),
-    help="Explicitly specify config file (it has to be a valid TOML file).",
-)
+@config_option
 def scaffold_(cfg):
-    """Manage folders needed for the project as spelled out in the configuration file."""
-    path = configs.detect(cfg)
-    base_configs = configs.load(path)
-    configs.configs = configs.defaults(base_configs)
-    if cfg is not None:
-        print(f"Configurations loaded from '{path}'")
+    """Detect amd load configuration file."""
+    load_config(cfg)
 
 
 @scaffold_.command()
